home/pyproject.toml
[tool.poetry]
name = "home"
version = "0.1.0"
description = "Manage the home automation"
authors = ["Dorian Jaminais <epa@jaminais.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
asyncio-mqtt = "^0.10.0"
argparse-logging = "^2020.11.26"
pytz = "^2021.3"
aiohttp = "^3.8.0"
urllib-ext = "^2021.10.26"
aioprometheus = {extras = ["starlette"], version = "^21.9.1"}
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
colorlog = "^6.6.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
mypy = "^0.910"
isort = "^5.10.0"
pytest = "^6.2.5"
types-pytz = "^2021.3.0"
pytest-asyncio = "^0.16.0"
ipython = "^7.29.0"
types-PyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
